Fix initial formating on table
Initially on the homepage table load (just the first time it loads, not on a new search), the spacing of the columns and the elements within the table are off.  This is fixed when the elements are moved around (sorted).  A virtualy click sent to the ID column after the table has loaded.