Support Muiltple File Paths
Moving away from the FilePath name in the db and implementing the following,

FilePathOriginal - The original err/out/etc. files
FilePathFilter - The filtered versions of the err/out/etc. files

The webserver will serve the FilePathOriginal version for now and will serve FilePathFilter in the future by default and allow the config file to choose which one to serve.